added authentication to journal endpoints

+++ journalApp/src/main/java/com/example/journalapp/services/JournalEntryService.java
@@ -28,7 +28,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntry().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }catch (Exception e) {
             System.out.print(e.getMessage());
             throw new RuntimeException("Something uncertain happened in JES saveEntry Method.");
@@ -47,10 +47,24 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findUser(userName);
-        user.getJournalEntry().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findUser(userName);
+            removed = user.getJournalEntry().removeIf(x -> x.getId().equals(id));
+            if(removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch (Exception e) {
+            System.out.println(e.getMessage());
+            throw new RuntimeException("Error occurred while deleting something.");
+        }
+        return removed;
+    }
+
+    public List<JournalEntry> findByUsername(String username) {
+        return null;
     }
 }
